refactor: useAnimatedModal -> useModal로 네이밍 수정 및 로직 간소화
